requirements.txt
streamlit
semantic-kernel
aiohttp
azure-identity
python-dotenv
tiktoken